refactor: move NBits hint to math/bits
