+++ backend/hint/builtin.go
@@ -11,20 +11,10 @@ var (
 	// corresponds to checking if a == 0 (for which the function returns 1) or a
 	// != 0 (for which the function returns 0).
 	IsZero = NewStaticHint(isZero)
-
-	// IthBit returns the i-tb bit the input. The function expects exactly two
-	// integer inputs i and n, takes the little-endian bit representation of n and
-	// returns its i-th bit.
-	IthBit = NewStaticHint(ithBit)
-
-	// NBits returns the n first bits of the input. Expects one argument: n.
-	NBits = NewStaticHint(nBits)
 )
 
 func init() {
 	Register(IsZero)
-	Register(IthBit)
-	Register(NBits)
 }
 
 func isZero(curveID ecc.ID, inputs []*big.Int, results []*big.Int) error {
@@ -45,24 +35,5 @@ func isZero(curveID ecc.ID, inputs []*big.Int, results []*big.Int) error {
 	inputs[0].SetUint64(1)
 	result.Sub(inputs[0], result).Mod(result, q)
 
-	return nil
-}
-
-func ithBit(_ ecc.ID, inputs []*big.Int, results []*big.Int) error {
-	result := results[0]
-	if !inputs[1].IsUint64() {
-		result.SetUint64(0)
-		return nil
-	}
-
-	result.SetUint64(uint64(inputs[0].Bit(int(inputs[1].Uint64()))))
-	return nil
-}
-
-func nBits(_ ecc.ID, inputs []*big.Int, results []*big.Int) error {
-	n := inputs[0]
-	for i := 0; i < len(results); i++ {
-		results[i].SetUint64(uint64(n.Bit(i)))
-	}
 	return nil
 }
